specify float32 in feature engine

+++ training/env/featureEngine.py
@@ -49,20 +49,20 @@ class FeatureEngineDummy(FeatureEngine):
 
 
 class FeatureEngineVersion1(FeatureEngine):
-    
+
     def get_input_shape(self):
         return 17
-    
+
     def get_feature(self, observation) -> torch.Tensor:
 
-        mid_price = (observation['ap0'] + observation['bp0']) / 2 
+        mid_price = (observation['ap0'] + observation['bp0']) / 2
 
         ask_price_levels = [
             observation['ap0'] / mid_price - 1,
             observation['ap1'] / mid_price - 1,
             observation['ap2'] / mid_price - 1,
             observation['ap3'] / mid_price - 1,
-            observation['ap4'] / mid_price - 1,  
+            observation['ap4'] / mid_price - 1,
         ]
         ask_vol_levels = [
             observation['av0'],
@@ -76,7 +76,7 @@ class FeatureEngineVersion1(FeatureEngine):
             observation['bp1'] / mid_price - 1,
             observation['bp2'] / mid_price - 1,
             observation['bp3'] / mid_price - 1,
-            observation['bp4'] / mid_price - 1,  
+            observation['bp4'] / mid_price - 1,
         ]
         bid_vol_levels = [
             observation['bv0'],
@@ -87,7 +87,7 @@ class FeatureEngineVersion1(FeatureEngine):
         ]
 
         avg_price_to_trade_list = [self.avg_price_to_trade(
-                        observation, 
+                        observation,
                         vol_to_trade,
                         ask_price_levels,
                         ask_vol_levels,
@@ -105,55 +105,57 @@ class FeatureEngineVersion1(FeatureEngine):
             self.relative_time(observation),
             self.price_log(observation),
             self.mid_price_relative(observation),
-            #*ask_price_levels,
-            #*bid_price_levels,
+            # *ask_price_levels,
+            # *bid_price_levels,
             *avg_price_to_trade_list,
-        ])
+        ],
+            dtype=torch.float32,
+        )
 
         return feature_tensor
 
     def relative_time(self, observation):
-        
+
         e_time = int(observation['eventTime']) // 1000
         hours = e_time // 10000
         minutes = (e_time // 100) % 100
         seconds = e_time % 100
-        
+
         std_time = hours * 3600 + minutes * 60 + seconds
         if e_time > 130000:
             std_time -= 5400
-        
+
         std_time -= 34200
 
         return std_time / 14400 * 2 - 1
 
-        
+
     def price_log(self, observation):
-        
+
         open_price = observation['ap0_t0']
         norm_price = np.log(open_price) / np.log(15000)
-        
+
         return norm_price * 2 - 1
-    
+
 
     def mid_price_relative(self, observation):
         return (observation['ap0'] + observation['bp0']) / ( 2 * observation['ap0_t0']) - 1
 
-    def avg_price_to_trade(self, 
-                           observation, 
+    def avg_price_to_trade(self,
+                           observation,
                            vol_to_trade: int,
                            ask_price_levels,
                            ask_vol_levels,
                            bid_price_levels,
                            bid_vol_levels,
                            ):
-        
+
         abs_vol = abs(vol_to_trade)
 
         if abs_vol > 10 or vol_to_trade == 0:
             raise ValueError(f"Feature Error|avg_price_to_trade|{vol_to_trade}|{observation}")
 
-        total_cost = 0        
+        total_cost = 0
         if vol_to_trade > 0:
             for ap, av in zip(ask_price_levels, ask_vol_levels):
                 if vol_to_trade > av:
@@ -171,11 +173,11 @@ class FeatureEngineVersion1(FeatureEngine):
                     total_cost += bv * bp
                 else:
                     total_cost += vol_to_trade * bp
-                    break        
+                    break
 
-        return total_cost / abs_vol   
+        return total_cost / abs_vol
 
 
-                    
+
